refactor: 패키지 구조 변경 및 H2Runner 클래스 제거
