--- src/main/java/antigravity/config/H2Runner.java
@@ -1,29 +1,0 @@
-package antigravity.config;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.ApplicationArguments;
-import org.springframework.boot.ApplicationRunner;
-import org.springframework.stereotype.Component;
-
-import javax.sql.DataSource;
-import java.sql.Connection;
-
-@RequiredArgsConstructor
-@Component
-public class H2Runner implements ApplicationRunner {
-
-    private final DataSource dataSource;
-
-    @Override
-    public void run(ApplicationArguments args) throws Exception {
-
-        try (Connection connection = dataSource.getConnection()) {
-            System.out.println(connection);
-            String URL = connection.getMetaData().getURL();
-            System.out.println(URL);
-            String User = connection.getMetaData().getUserName();
-            System.out.println(User);
-        }
-    }
-
-}
